Control.Monad.Heap

Description

The Heap monad: a monad for efficient weighted search.

This module provides an implementation of the Heap monad transformer as described in:

Donnacha Oisín Kidney and Nicolas Wu. 2021. Algebras for weighted search. Proc. ACM Program. Lang. 5, ICFP, Article 72 (August 2021), 30 pages. DOI:https://doi.org/10.1145/3473577

This monad transformer can be used to implement search algorithms like Dijkstra's algorithm (see MonusWeightedSearch.Examples.Dijkstra), or the Viterbi algorithm (MonusWeightedSearch.Examples.Viterbi), or probabilistic parsing (MonusWeightedSearch.Examples.Parsing).

The type supports nondeterminism (using the Alternative and MonadPlus interfaces), where each branch in a computation can be weighted by some Monus. A Monus is an ordered Monoid with some pseudo-subtraction operator, see the Data.Monus module for more details.

Heap Type

newtype HeapT w m a

The HeapT monad transformer: a monad for weighted search.

This monad supports nondeterminism through the Alternative and MonadPlus classes, but different branches in the computation may be weighted by the w parameter. A computation can be given a specific weight using the MonadWriter interface:

  tell 4 >> xs

This weights the computation xs with 4.

Depending on the Monus used, the order of the search can be specified. For instance, using the Monus in Data.Monus.Dist, we have the following:

>>> search (fromList [('a',5), ('b', 3), ('c',6)])
[('b',3),('a',5),('c',6)]

>>> search (fromList [('b',3), ('a',5), ('c',6)])
[('b',3),('a',5),('c',6)]

Constructors

HeapT

Fields

runHeapT :: ListT m (Node w a (HeapT w m a))

data Node w a b

A Heap is a list of Nodes of Heaps.

Constructors

Leaf a

!w :< b infixr 5

Non-transformer form

type Heap w = HeapT w Identity

The Heap type, specialised to the Identity monad.

pattern Heap :: [Node w a (Heap w a)] -> Heap w a

The constructor for the non-transformer Heap type.

runHeap :: Heap w a -> [Node w a (Heap w a)]

Constructing Heaps

fromList :: Applicative m => [(a, w)] -> HeapT w m a

Build a heap from a list of values paired with their weights.

Popping the smallest element

popMin :: Monus w => Heap w a -> ([a], Maybe (w, Heap w a))

O(log n). popMin returns a list of those elements in the Heap with a weight equal to mempty, paired with the rest of the heap and the minimum weight in the rest of the heap.

popMinT :: (Monus w, Monad m) => HeapT w m a -> m ([a], Maybe (w, HeapT w m a))

The monadic variant of popMin.

Turning into a cons-list

flatten :: Monus w => Heap w a -> [(w, [a])]

O(n log n). Return all the elements of the heap, in order of their weights, grouped by equal weights, paired with the differences in weights.

The weights returned are the differences, not the absolute weights.

>>> flatten (fromList [('a',5), ('b', 3), ('c',6)])
[(0,""),(3,"b"),(2,"a"),(1,"c")]

flattenT :: (Monad m, Monus w) => HeapT w m a -> ListT m (w, [a])

The monadic version of flatten.

Searching the whole heap

search :: Monus w => Heap w a -> [(a, w)]

O(n log n). Return all of the elements in the heap, in order, paired with their weights.

searchT :: (Monad m, Monus w) => HeapT w m a -> m [(a, w)]

The monadic variant of search.

Returning one element

best :: Monus w => Heap w a -> Maybe (w, a)

O(log n). Return the lowest-weight element in the heap, paired with its weight.

bestT :: (Monad m, Monus w) => HeapT w m a -> m (Maybe (w, a))

The monadic variant of best.