Control.Monad.Heap.List

Description

A simple re-implementation of the list monad transformer, for use by the heap monad.

Type Definition

newtype ListT m a

A list monad transformer, "done right".

Constructors

ListT

Fields

runListT :: m (ListCons a (ListT m a))

data ListCons a b

The list constructor.

Constructors

Nil

a :- !b infixr 5

Building the list

unfoldrM :: Functor m => (b -> m (Maybe (a, b))) -> b -> ListT m a

Unfold, monadically, a list from a seed.

Running the list

toListT :: Monad m => ListT m a -> m [a]

Flatten all of the effects in the list and collect the results.

Transforming lists

catMaybesT :: Monad m => (a -> Maybe b) -> ListT m a -> ListT m b

Filter the list. An analogue of mapMaybe on lists.

listMmap :: Functor m => (m (ListCons a (ListT n b)) -> n (ListCons b (ListT n b))) -> ListT m a -> ListT n b

Apply a function to every effect layered in the list transformer.