Data.Monus.Dist

Description

A Monus for discrete distances.

Documentation

newtype Dist

A very simple Monus, based on the addition Monoid on Natural numbers. This represents discrete distances.

Constructors

Dist

Fields

runDist :: Natural

type Graph a = a -> [(a, Dist)]

A simple graph with Dist-weighted edges.

Note that the algorithms in this package can use any monus, not just Dist: we specialise here just for simplicity of presentation.