Data.Monus.Prob

Description

A Monus for probability.

Documentation

newtype Prob

A Monus for probabilities, where the underlying Monoid is the product monoid.

NB: The order on this type is reversed from the "usual" order on probability. i.e.

>>> 0.8 < (0.4 :: Prob)
True

Constructors

Prob

Fields

runProb :: Ratio Natural