MonusWeightedSearch.Examples.PairingFold

Description

This module explores different ways to implement the "pairing fold".

Documentation

>>> :{
data Tree = Leaf | Tree :*: Tree
instance Show Tree where
  show Leaf = "."
  show (xs :*: ys) = "(" ++ show xs ++ "*" ++ show ys ++ ")"
:}

pairFold1 :: (a -> a -> a) -> [a] -> Maybe a

The traditional definition of the pairing fold, as given in the original paper.

>>> pairFold1 (:*:) (replicate 5 Leaf)
Just ((.*.)*((.*.)*.))

pairFold2 :: (a -> a -> a) -> [a] -> Maybe a

data Acc3 a

Constructors

Acc3 (Maybe a) (Maybe a)

pairFold3 :: (a -> a -> a) -> [a] -> Maybe a