MonusWeightedSearch.Examples.Sampling

Description

Random sampling from the Heap monad.

The Heap monad can function as a probability monad, and it implements an efficient sampling algorithm, based on reservoir sampling.

Documentation

sample :: Heap Prob a -> IO a

Sample a single value from the heap.