MonusWeightedSearch.Examples.SubsetSum

Description

An implementation of shortest subset sum (the Inclusion-Exclusion method) using the Heap monad.

Documentation

inclusion :: Monad m => HeapT Dist m Bool

A weight for the inclusion or exclusion of an element.

This lets us weight the computation by number of elements included, and therefore optimise for the fewest.

shortest :: Int -> [Int] -> [Int]

shortest n xs returns the shortes subset of xs which sums to n.

>>> shortest 5 [10,-4,3,11,6,12,1]
[-4,3,6]