MonusWeightedSearch.Internal.AdjList

Description

This module provides an implementation of weighted graphs as arrays of words. It's useful for generating random graphs, and performance testing.

Documentation

newtype AdjList

A graph with vertices labelled by words, and edges weighted by words.

Constructors

AdjList

Fields

edges :: UArray (Word, Word) Word

randAdjList

Arguments

:: Word
Edge Density (as a percentage)

-> Word
Size

-> IO AdjList

Generate a random graph.

alSize :: AdjList -> Word

Return the number of vertices in the graph.

toGraph :: AdjList -> Graph Word

Convert an adjacency list to a standard graph.

fromGraph

Arguments

:: Word
The number of vertices in the graph.

-> Graph Word
The graph.

-> AdjList

Convert a graph to an adjacency list.