MonusWeightedSearch.Internal.Heap

Description

A reference implementation of a pairing heap, to compare to the heap monad.

Documentation

data Heap a b

A pairing heap.

This implementation does use a monus rather than just a standard ordered key, but that does not change any of the algorithms really.

Constructors

Leaf

Node !a b [Heap a b]

minView :: Monus a => Heap a b -> Maybe ((a, b), Heap a b)

O(log n). Pop the minimum element and its key in the heap, and return it.

singleton :: a -> b -> Heap a b

A singleton heap.

dijkstra :: Ord a => Graph a -> Graph a

An implementation of Dijkstra's algorithm on Graphs.

monusSort :: [Dist] -> [Dist]

Sort a list of Dist.