MonusWeightedSearch.Internal.TestHelpers

Description

Some utility functions for implementing Arbitrary etc.

Documentation

>>> import Test.QuickCheck

sumsTo :: Int -> Gen [Int]

sumsTo n generates a list that sums to n.

n >= 0 ==> forAll (sumsTo n) (\xs -> sum xs === n)

percentageChance :: Word -> Gen Bool

percentageChance n is True n% of the time, and False the rest of the time.

percentageChanceIO :: Word -> IO Bool

percentageChanceIO n is True n% of the time, and False the rest of the time.