MonusWeightedSearch.Internal.Heap

Description

A reference implementation of a pairing heap, to compare to the heap monad.

Documentation

type Heap a b = Maybe (Root a b)

A pairing heap.

This implementation does use a monus rather than just a standard ordered key, but that does not change any of the algorithms really.

data Root a b

A non-empty heap.

Constructors

Node !a b [Root a b]

minView :: Monus a => Heap a b -> Maybe ((a, b), Heap a b)

singleton :: a -> b -> Heap a b

A singleton heap.

dijkstra :: Ord a => Graph a -> Graph a

An implementation of Dijkstra's algorithm on Graphs.

monusSort :: Ord a => [a] -> [a]

Heapsort.